backend/prosto_board_volume-main/utils/pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.1.0",
    "requests>=2.32.3",
]
